requirements.txt
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0